+++ lib/alias.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/widgets.dart';
 
 typedef MAAlign = MainAxisAlignment;
+typedef MASize = MainAxisSize;
 typedef CAAlign = CrossAxisAlignment;