+++ lib/widget/align.dart
@@ -1,6 +1,7 @@
-import 'package:flutter/cupertino.dart';
+import 'package:flutter/widgets.dart';
 
 extension AlignWidgetX on Widget {
+  /// see [Align]
   Align align({
     Key? key,
     Alignment at = Alignment.center,