+++ lib/widget/center.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
 
 extension WidgetCenterX on Widget {
+  /// see [Center]
   Center center({Key? key, double? wFactor, double? hFactor}) => Center(
         key: key,
         widthFactor: wFactor,
@@ -8,6 +9,7 @@ extension WidgetCenterX on Widget {
         child: this,
       );
 
+  /// see [Center]
   Center centerAll(double factor, {Key? key}) => Center(
         key: key,
         widthFactor: factor,