+++ lib/widget/clip_r_rect.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
 
 extension ClipRRectWidgetX on Widget {
+  /// see [ClipRRect]
   ClipRRect clipRRect({
     Key? key,
     BorderRadiusGeometry? borderRadius,