+++ lib/widget/container.dart
@@ -1,5 +1,7 @@
-import 'package:flutter/cupertino.dart';
+import 'package:flutter/widgets.dart';
+
 extension ContainerWidgetX on Widget {
+  /// see [Container]
   Container container({
     Key? key,
     AlignmentGeometry? align,