+++ lib/widget/scroll_view.dart
@@ -2,6 +2,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/widgets.dart';
 
 extension SingleChildScrollViewWidgetX on Widget {
+  /// see [SingleChildScrollView]
   SingleChildScrollView scrolled({
     Key? key,
     Axis scrollDirection = Axis.vertical,