+++ lib/widget/sizedbox.dart
@@ -1,12 +1,14 @@
 import 'package:flutter/cupertino.dart';
 
 extension WidgetSizedBox on Widget {
+  /// see [SizedBox]
   SizedBox sized({Key? key, double? w, double? h}) => SizedBox(
         key: key,
         width: w,
         height: h,
         child: this,
       );
+  /// see [SizedBox]
   SizedBox sizedIn(Size size,{Key? key}) => SizedBox(
         key: key,
         width: size.width,
@@ -14,6 +16,7 @@ extension WidgetSizedBox on Widget {
         child: this,
       );
 
+  /// see [SizedBox]
   SizedBox sizedAll(double length, {Key? key}) => SizedBox(
         key: key,
         width: length,
@@ -21,6 +24,7 @@ extension WidgetSizedBox on Widget {
         child: this,
       );
 
+  /// see [SizedBox]
   SizedBox shrinkSize({Key? key}) => SizedBox.shrink(
         key: key,
         child: this,